chore: rename core modules to snake_case (match bundle-examples)
Mirror the module rename from databricks/bundle-examples PR #163
(commit d31b7196, "Address PR review feedback: snake_case modules"):

- DbtFactory.py   -> dbt_factory.py
- DbtTask.py      -> dbt_task.py
- TaskFactory.py  -> task_factory.py
- Utils.py        -> utils.py
- __about__.py    -> __version__.py

Only module *paths* change (imports, the `[tool.hatch.version]` path); the
public class names (`DbtFactory`, `DbtTask`, `TaskFactory`, `TestTaskFactory`,
`DbtTaskOptions`, ...) are unchanged, matching bundle-examples.

Branched off main after the `ruff format` migration (#32) and config cleanup
(#33) merged. With that formatting already in place, the renamed core
(`dbt_factory.py`, `dbt_task.py`, `task_factory.py`, `utils.py`,
`__version__.py`, `notebook/run_dbt_command.py`) is byte-for-byte identical to
bundle-examples' vendored copy at PR #163 head — verified with `diff`.

The release path is unaffected: `[tool.hatch.version] path` now points at
`__version__.py`; `hatch build` produces databricks_dbt_factory-0.3.3 with the
version and entry point resolving.

`make lint` and `make test` (317 passed) are green.
